Fix unwanted padding and a visual jump in the banner change description.
When setting the banner change description, the field performs a little
jump, first shrinking to accomodate all text and then expanding
slightly. This is because the margin on a part of the banner was being
applied to the text, creating a little bit of extra space at the bottom
of the field.

This change tightens up the CSS to only apply the styling to the correct
parts of the banner, fixing this visual issue.

Testing Done:
Tested clearing the field, saving, entering new text, and saving. There
was no longer a visual jump in the field. It instead maintained the
correct size.

Reviewed at https://reviews.reviewboard.org/r/13287/