Remove the Pygments error marker for syntax-highlighted diffs.
Pygments doesn't always recognize all syntax in a file, and while this
is often harmless, it comes with a lot of visual noise in the form of a
red border around the unexpected parts.

The indicator was more useful once upon a time, but with modern
practices around lint checking and CI builds, these markers are more
annoying than helpful.

This change removes visual indicators on unknown syntax, allowing
reviewers to concentrate more on the code without the distractions.

Testing Done:
Posted a change with a LessCSS file that triggered Pygments error
markers. Saw that the change lacked the borders.

Reviewed at https://reviews.reviewboard.org/r/13290/