Update the repository list to show SCMTool information.
The repository list in the admin UI previously showed information on the
account and hosting service associated with a repository, but there was
nothing useful for non-hosted repositories.

This has now been updated to show the SCMTool type for plain
repositories, and to avoid any problems if failing to load the SCMTool.

The repository type is also shown for hosting services, since they may
support multiple types of repositories.

To make the purpose of the column more clear, this has been renamed from
"Hosting Service Account" to "Type / Account".

Testing Done:
Viewed my list of repositories and saw the repository types for all
repositories.

Reviewed at https://reviews.reviewboard.org/r/12984/