Hide the Commits field when there are no commits present.
The Commits field on the main review request box is present even if
there's no commits data to show. We end up with a label but no content.
This happens because we specify the field should be always present
unless this is on the diff viewer page (which has its own commits
representation).

We now factor in the value itself, ensuring that we only show the field
if we know we're going to end up showing a list of commits.

Testing Done:
Tested viewing review requests with and without commits data. I could
no longer reproduce the bug where only a Commits label was shown.

Reviewed at https://reviews.reviewboard.org/r/13296/