Enable HTTP/2 support for Apache when available.
This updates our default Apache configuration template to opt into
HTTP/2 when available on the server, which will help bring better
performance for new installs. While it won't affect existing installs,
the documentation has been updated to include this as well.

Apache accepts this protocol as a request, and makes a determination as
to whether it can be supported based on module availability and SSL.

Testing Done:
Installed a new site and saw the new configuration.

Tested this on a server.

Reviewed at https://reviews.reviewboard.org/r/13301/