Update and improve docs for new changes in RB6.
We have a lot of new UI and user experience changes in RB6, so it was
necessary to do an overview of the docs to make sure everything is up to date.
This change updates any screenshots that were affected by our UI changes, and
fleshes out some documentation for file attachments and review requests.

Note: The screenshots in this change are not perfect and are just
      placeholders. They should be replaced with new screenshots in a later
      change.

Testing Done:
Built the docs, made sure the new screenshots looked good, checked that all new
refs worked, and checked for spelling errors.

Reviewed at https://reviews.reviewboard.org/r/13299/