Fix handling URLErrors in GitHub without an HTTP status code.
The HTTP error handler in GitHub was assuming that error classes always
had an HTTP status code, but this is only the case for `HTTPError` and
not `URLError`. If an error occurs due to an outage or something other
than an HTTP response, we'd crash.

This fixes the code to provide an HTTP status code only if it exists on
the error object.

Testing Done:
Turned off wifi and attempted to load a diff locally. Saw an error about
issues accessing the file, rather than a crash traceback.

Reviewed at https://reviews.reviewboard.org/r/13303/