Fix the display of buttons and the login page on mobile.
The login page had a broken look on mobile due to a missing selector,
causing some of our style resets for labels to fail to apply. This adds
the missing selector, fixing up the field layout.

The recent changes to button layout also impacted this page, and others
in the UI. The usage of `display: inline-flex` caused the text to be
left-aligned. This impacted all buttons in the UI with a non-default
width. We now center-justify the content of buttons to restore the
original centering.

Testing Done:
Tested the login page on desktop and mobile.

Checked other buttons throughout the UI for regressions.

Reviewed at https://reviews.reviewboard.org/r/13310/