Fix a regression in building the documentation.
With the introduction of `get_data_dir()`, documentation was failing to
build, due to the lack of the expected data directory.

This change fixes this by simply adding the directory, along with a
`.gitignore` that ignores everything inside of it (as files may be
generated during a docs run).

Testing Done:
Successfully built the docs.

Reviewed at https://reviews.reviewboard.org/r/13308/