Fix the display of the comment dialog on mobile devices.
This updates the comment dialog to have a maximum width that fits
cleanly within the browser's viewport, with some spacing on the side.
This is there for mobile, where the comment dialog's default width would
easily overflow the page.

With this change, invoking the comment dialog will now properly display
it within the bounds of the viewport, allowing the user to leave a
comment and then save it without having to work around the viewport or
size.

Testing Done:
Tested on desktop, responsive mode on Chrome and Firefox, and on iOS 17.

Reviewed at https://reviews.reviewboard.org/r/13311/