Add DiffCommit.commit_message_body for getting the body of a commit.
This property returns the commit message's body, which is the part after
the summary, if it's available. It will be used as part of a fix for the
diff commit list.

Unit tests are added. The old test suite had setup code that was very
specific to the tests that were initially written, so some cleanup was
done to move that setup code into a dedicated, optional function.

Testing Done:
Unit tests pass.

Made use of this function in some upcoming code.

Reviewed at https://reviews.reviewboard.org/r/13332/