Add RB.DiffCommit.commitMessageBody in JavaScript.
This adds `RB.DiffCommit.commitMessageBody`, which contains the part of
the commit message that follows the summary. This allows for processing
the summary and the body separately, which will be needed for an
upcoming change.

Testing Done:
JavaScript unit tests pass.

Made use of this in an upcoming change.

Reviewed at https://reviews.reviewboard.org/r/13333/