Fix a regression in loading fields in the issue summary table.
The recent fields layout update broke a selector for the issue summary
table. Some assumptions were made about the placement of the
`.rb-c-issue-summary-table` element, which broke. This in turn broke the
filters, tabs, and other behavior.

This change fixes the selector, restoring all functionality, and adding
an assertion so this is caught in the future.

Testing Done:
Tested behavior on desktop and mobile. Verified I could filter by users
and by status again.

Reviewed at https://reviews.reviewboard.org/r/13338/