Fix diff API security issues on Review Board 5.
During an internal audit of our API, we found that it was possible for a
user to construct an HTTP GET request to the diff and filediff APIs to
retrieve patch and serialized diff data content when the user lacked
permission to access the review request, such as when the review request
is in a draft or on a private repository. This was due to missing access
checks that were normally performed by the core API functionality but
were skipped when specialized for handling this form of request.

We are not aware of any incidents regarding this issue, but we consider
it a severe issue for companies utilizing private repositories and
recommend that all server administrators upgrade immediately.

We have enhanced our test suite to check for these issues, have
completed an audit of all the API resources we provide, and are in the
process of reworking how our APIs are implemented so that access checks
are handled by the base API layer and not per-API.

Releases are being made for Review Board 3, 4, 5, and 6 to cover this
issue.