Clean up the code for the datagrid views.
The code for datagrid views was pretty ancient, poorly-documented, and
lacked a lot of best practices and modern standards. While not a blocker
for anything, given that some work is going to be done in this module,
it felt time to clean some of this up.

This change updates the module for type checking, modern documentation
standards, and to remove unnecessary arguments to views and to construct
datagrids and query functions with keyword arguments. This will provide
a good base for some further updates.

Testing Done:
Unit tests passed.

Tested all the datagrids manually.

Reviewed at https://reviews.reviewboard.org/r/13354/