Prepare ReviewManager tests for upcoming query improvements.
This reworks the code in `test_review_manager.py` to allow more of the
queries and the test setup to be shared. We now have tests for the
global site, Local Sites, and the global site when there are Local Sites
in the database.

Upcoming work on review queries will take advantage of this to check
queries based on the various Local Site conditions.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13356/