Update review queries to query Local Sites conditionally.
Review Board 5 introduced some mechanisms for optimizing queries to
avoid filering by Local Sites if there aren't any Local Sites in the
database. We updated a few queries to use this, but not all.

This change updates `ReviewManager` to take advantage of this, reducing
the complexity of queries and helping the database better use its
indexes on a typical install.

There's also a fix for building queries for Local Sites, taking
advantage of the field name provided.

This is the first in a series of changes for optimizing Local Site
filters out of queries.

Testing Done:
All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13357/