Fix batch endpoint URLs when running with a SITE_ROOT.
Our new implementation of the publish and dashboard close/archive operations
use a special endpoint to perform batch operations on groups of review
requests and reviews. The JS side of this unfortunately did not include
the site root in the URL, breaking for people who use this type of
deployment. This change adds that to the URLs.

Testing Done:
- Ran unit tests.
- Verified that I could publish things.

Reviewed at https://reviews.reviewboard.org/r/13385/