Fix and add missing installation steps in the documentation.
We had a few issues that were noted while going through an installation
process recently. There were a couple places where we specified the
wrong filesystem path to `pip`, and where we didn't document some
packages for Apache or apt build sources that may be needed if not
already installed on the system.

This change fixes up the documentation and adds the steps for installing
Apache + mod_wsgi, and enabling build sources. It also changes the order
of packages for the initial system package installation steps to help with
scanning and updating package entries.

Testing Done:
Built the docs and read through them.

Ran the steps on an Ubuntu 22.04 system.

Reviewed at https://reviews.reviewboard.org/r/13383/