Expand the Local Site query checks in the Users datagrid tests.
This splits most of the unit tests in the Users datagrid test suite into
a global site test, a Local Site test, and a global site + Local Sites
in the database test.

This gives us thorough query coverage, which unveils a number of
sub-optimal queries that will need to be addressed in upcoming changes.

The queries are intentionally verbose right now, with no query-building
being shared across tests, and with the query-by-query breakdowns
removed. This is intentional, as upcoming work is expected to heavily
impact these queries.

There's also a little bit of base class test setup, teardown, and
utilities that are introduced that will be needed for the test query
work.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13360/